[lldb] Convert LocateSymbolFile into a plugin
This commit contains the initial scaffolding to convert the
functionality currently implemented in LocateSymbolFile to a plugin
architecture. The plugin approach allows us to easily add new ways to
find plugins.

The current implementation abuses the host OS to use the DebugSymbols
framework on macOS. With the plugin approach, we retain all the benefits
(including the ability to compile it out) with a much cleaner level of
abstraction.

To limit the scope of this patch, I've only converted
LocateExecutableObjectFile. The goal is to do the same for the remaining
LocateSymbolFile functions and eventually remove the file. To make
reviewing easier, that will done as follow-up patches.