[lldb-dap] Rework attach tests
I traced the SIGHUP I was receiving (#133782) when running the lldb-dap
tests to `TestDAP_attach.py`. At first I thought the issue was caused by
the Python logic, but I was able to reproduce the issue no matter which
test function I skipped, leading me to the test inferior.

Pavel improved the test a while ago by using file-based synchronization,
but the test was still sleeping for 10 seconds.